Fix sign-extension issue in VR4300_SHRL.

+++ vr4300/functions.c
@@ -475,7 +475,7 @@ void VR4300_SRL(struct vr4300 *vr4300, uint64_t unused(rs), uint64_t rt) {
   unsigned dest = GET_RD(iw);
   unsigned sa = iw >> 6 & 0x1F;
 
-  exdc_latch->result = (int32_t) (rt >> sa);
+  exdc_latch->result = (int32_t) ((uint32_t) rt >> sa);
   exdc_latch->dest = dest;
 }
